Page Fault Vector Typo (#123)
* Page Fault Vector Typo

* Femtoseconds conversion

* RSDP Typo 

The signature for the RSDP is "RSD PTR ", and it is not an SDT. 

https://uefi.org/specs/ACPI/6.5/05_ACPI_Software_Programming_Model.html?highlight=rsdt#root-system-description-table-rsdt

* APIC Address Incorrect

Fixed the description of the APIC address (the physaddr is not shifted, but the lower bits are masked), and fixed a typo with the location. On my ``qemu-system-x86_64``, the LAPIC base is 0xFEE00000 in physical memory.

* Latex-style exponent

* C struct syntax error

You need to create a entry in the symbol table to reference a C-object itself, e.g. for self-referencing in a struct